fixed money formatting precision; fixed prepack summary output ratio; moved position summary folder to common from recipe;
