+++ apps/cookbook-mobile/src/ui/common/summary/position-summary/product-position-summary.tsx
@@ -2,8 +2,8 @@ import { ProductIngredient } from "@cookbook/domain/types/position/product-ingre
 import { FormatNumber } from "@cookbook/domain/util";
 import { t } from "i18next";
 import { View } from "react-native";
-import { DividedRow, PositionRowLabel, isServedInUnits } from "../../../../common/summary/label-components";
-import { styles } from "../recipe-summary.style";
+import { DividedRow, PositionRowLabel, isServedInUnits } from "../label-components";
+import { styles } from "../../../home/recipes/recipe-summary/recipe-summary.style";
 
 export function ProductPositionSummary(one: ProductIngredient, ratio: number) {
     return <DividedRow>
